+++ npf/cmdline.py
@@ -1,32 +1,6 @@
-import sys
-import argparse
-import os
-
-from pathlib import Path
 from argparse import ArgumentParser
-from typing import Dict, List
+import argparse
 
-import regex
-import re
-from decimal import Decimal
-
-from npf.node import Node
-
-import numpy as np
-
-options = None
-cwd = None
-
-def get_valid_filename(s):
-    s = str(s).strip().replace(' ', '_')
-    return re.sub(r'(?u)[^-\w.]', '', s)
-
-class ExtendAction(argparse.Action):
-    def __init__(self, option_strings, dest, nargs=None, **kwargs):
-        super(ExtendAction, self).__init__(option_strings, dest, nargs, **kwargs)
-
-    def __call__(self, parser, namespace, values, option_string=None):
-        setattr(namespace, self.dest, getattr(namespace, self.dest) + values)
 
 
 def add_verbosity_options(parser: ArgumentParser):
@@ -128,10 +102,20 @@ def add_graph_options(parser: ArgumentParser):
                    default=False)
     s.add_argument('--statistics-maxdepth',
                    help='Max depth of learning tree', dest='statistics_maxdepth', type=int, default=None)
+    s.add_argument('--statistics-maxmetrics',
+                   help='Max depth of learning tree', dest='statistics_maxmetrics', type=int, default=3)
     s.add_argument('--statistics-filename',
                    help='Output of learning tree', dest='statistics_filename', type=str, default=None)
 
     return g
+
+
+class ExtendAction(argparse.Action):
+    def __init__(self, option_strings, dest, nargs=None, **kwargs):
+        super(ExtendAction, self).__init__(option_strings, dest, nargs, **kwargs)
+
+    def __call__(self, parser, namespace, values, option_string=None):
+        setattr(namespace, self.dest, getattr(namespace, self.dest) + values)
 
 
 def add_testing_options(parser: ArgumentParser, regression: bool = False):
@@ -203,7 +187,6 @@ def add_testing_options(parser: ArgumentParser, regression: bool = False):
     return t
 
 
-
 def add_building_options(parser):
     b = parser.add_argument_group('Building options')
     bf = b.add_mutually_exclusive_group()
@@ -229,314 +212,3 @@ def add_building_options(parser):
                     help='Force to rebuild some dependencies', dest='force_build_deps',
                    action=ExtendAction, default=[], nargs='+')
     return b
-
-nodePattern = regex.compile(
-    "(?P<role>[a-zA-Z0-9]+)=(:?(?P<user>[a-zA-Z0-9]+)@)?(?P<addr>[a-zA-Z0-9.-]+)(:?[:](?P<path>[a-zA-Z0-9_./~-]+))?")
-roles = {}
-
-
-def nodes_for_role(role, self_role=None, self_node=None, default_role_map={}):
-    if role is None or role == '':
-        role = 'default'
-    if role == 'self':
-        if self_role:
-            role = self_role
-            if self_node:
-                return [self_node]
-        else:
-            raise Exception("Using self without a role context. Usually, this happens when self is used in a %file")
-    if role not in roles:
-        if role in default_role_map:
-            role = default_role_map[role]
-    return roles.get(role, roles['default'])
-
-
-def executor(role, default_role_map):
-    """
-    Return the executor for a given role as associated by the cluster configuration
-    :param role: A role name
-    :return: The executor
-    """
-    return nodes_for_role(role, default_role_map)[0].executor
-
-
-def set_args(args):
-    sys.modules[__name__].options = args
-    args.cwd = os.getcwd()
-
-def initialize(args):
-    set_args(args)
-
-    #other random stuffs to do
-    if not options.build_folder is None:
-        if not os.access(options.build_folder, os.W_OK):
-            raise Exception("The provided build path is not writeable or does not exists : %s!" % options.build_folder)
-        options._build_path = options.build_folder
-    else:
-        options._build_path = npf_writeable_root_path()+'/build/'
-
-    if type(options.use_last) is not int:
-        if options.use_last:
-            options.use_last = 100
-
-    if not os.path.exists(experiment_path()):
-        raise Exception("The experiment root '%s' is not accessible ! Please explicitely define it with --experiment-path, and ensure that directory is writable !" % experiment_path())
-
-    if not os.path.isabs(options.experiment_folder):
-        options.experiment_folder = os.path.abspath(options.experiment_folder)
-
-
-    options.search_path = set(options.search_path)
-    for t in [options.test_files]:
-        options.search_path.add(os.path.dirname(t))
-
-def create_local():
-    # Create the test file
-    os.close(os.open(experiment_path() + ".access_test" , os.O_CREAT))
-    local = Node.makeLocal(options)
-    #Delete the test file
-    os.unlink(experiment_path() + ".access_test")
-    roles['default'] = [local]
-    return local
-
-def parse_nodes(args):
-    initialize(args)
-    local = create_local()
-
-    for val in options.cluster:
-
-        # Create the test file
-        os.close(os.open(experiment_path() + ".access_test" , os.O_CREAT))
-
-        variables : list[str] = val.split(',')
-        if len(variables) == 0:
-            raise Exception("Bad definition of cluster parameter : %s" % variables)
-        mapping: str =variables[0].strip()
-        match = nodePattern.match(mapping)
-        if not match:
-            raise Exception("Bad definition of node : %s" % mapping)
-
-        path = match.group('path')
-
-        del variables[0]
-
-        nfs = None
-        assert isinstance(variables, list)
-        for opts in variables:
-            assert isinstance(opts, str)
-            var,val = opts.split('=')
-            if var == "nfs":
-                nfs = int(val)
-            elif var == "path":
-                path = val
-            else:
-                continue
-            variables.remove(opts)
-
-        if match.group('addr') == 'localhost':
-            node = local
-        else:
-            node = Node.makeSSH(user=match.group('user'), addr=match.group('addr'), path=path,
-                            options=options, nfs=nfs)
-        role = match.group('role')
-        if role in roles:
-            roles[role].append(node)
-            print("Role %s has multiple nodes. The role will be executed by multiple machines. If this is not intended, fix your --cluster option." % role)
-        else:
-            roles[role] = [node]
-
-        for opts in variables:
-            var,val = opts.split('=')
-            if var == 'nic':
-                node.active_nics = [ int(v) for v in val.split('+') ]
-            elif var == "multi":
-                node.multi = int(val)
-            elif var == "mode":
-                node.mode = val
-            else:
-                raise Exception("Unknown cluster variable : %s" % var)
-
-        #Delete the test file if it still exists (if the remote is the local machine, it won't)
-        if os.path.exists(experiment_path() + ".access_test"):
-            os.unlink(experiment_path() + ".access_test")
-
-def parse_variables(args_variables, tags, sec) -> Dict:
-    variables = {}
-    for variable in args_variables:
-        var, val, assign = sec.parse_variable(variable,tags)
-        if var:
-            val.assign = assign
-            variables[var] = val
-    return variables
-
-
-def override(args, tests):
-    for test in tests:
-        overriden_variables = parse_variables(args.variables, test.tags, test.variables)
-        overriden_config = parse_variables(args.config, test.tags, test.config)
-        test.variables.override_all(overriden_variables)
-        test.config.override_all(overriden_config)
-    return tests
-
-
-def npf_root_path():
-    # Return the path to NPF root
-    return os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
-
-def npf_writeable_root_path():
-    path = npf_root_path()
-    if not os.access(path, os.W_OK):
-        return experiment_path()
-    else:
-        return path
-
-def experiment_path(options = None):
-    # Return the path to NPF experiment folder
-    if options is None:
-        options = sys.modules[__name__].options
-    return os.path.join(options.experiment_folder,'')
-
-def cwd_path(options=None):
-    # Return the path to where NPF was first executed
-    if options is None:
-        options = sys.modules[__name__].options
-    return options.cwd
-
-def get_build_path(options = None):
-    if options is None:
-        options = sys.modules[__name__].options
-    assert(options._build_path)
-    return options._build_path
-
-def from_experiment_path(path):
-    # Returns the path under NPF root if it is not absolute
-    if (os.path.isabs(path)):
-        return path
-    else:
-        return (experiment_path() if os.path.isabs(experiment_path()) else os.path.abspath(experiment_path())) + os.sep + path
-
-def find_local(path, critical: bool = False, suppl: List = None):
-    if os.path.exists(path):
-        return path
-
-    searched = [npf_root_path(), '.', experiment_path()] + list(sys.modules[__name__].options.search_path) + (suppl if suppl else [])
-    for a in searched:
-        p = a + os.sep + path
-        if os.path.exists(p):
-            return p
-    if critical:
-        raise FileNotFoundError("Could not find file %s, locations searched :\n%s" %
-                (path,
-                    "\n".join(searched)))
-    return path
-
-def splitpath(hint):
-    if hint is None:
-        hint = "results"
-    dirname, c_filename = os.path.split(hint)
-    if c_filename == '':
-        basename = ''
-        ext = ''
-    else:
-        basename, ext = os.path.splitext(c_filename)
-        if not ext and basename.startswith('.'):
-            ext = basename
-            basename = ''
-    return dirname, basename, ext
-
-def build_filename(test, build, hint, variables, def_ext, type_str='', show_serie=False, suffix='', force_ext = False, data_folder = False, prefix=None):
-    var_str = get_valid_filename('_'.join(
-        ["%s=%s" % (k, (val[1] if type(val) is tuple else val)) for k, val in sorted(variables.items()) if val]))
-
-    if hint is None:
-        if data_folder:
-            path = build.result_path(test.filename, def_ext, folder = var_str + (('-' if var_str else '') + type_str if type_str else ''), prefix=prefix, suffix = ('-' + suffix if suffix else '') + ('-' + get_valid_filename(build.pretty_name()) if show_serie else ''))
-        else:
-            path = build.result_path(test.filename, def_ext, suffix=('-' + suffix if suffix else '') + var_str + ('-' + type_str if type_str else '') + ('-' + get_valid_filename(build.pretty_name()) if show_serie else '') , prefix=prefix)
-    else:
-        dirname, basename, ext = splitpath(hint)
-
-        if ext is None or ext == '' or force_ext:
-            ext = '.' + def_ext
-
-        if basename is None or basename == '':
-            basename = var_str
-
-        if not data_folder:
-            if prefix:
-                basename = prefix + basename
-
-            if not dirname or show_serie:
-                basename = (get_valid_filename(build.pretty_name()) if show_serie else '') + basename
-            path = (dirname + '/' if dirname else '') + basename + (
-        ('-' if basename else '') + type_str if type_str else '') + ('' if not suffix else ("-" + suffix)) + ext
-        else:
-            if not dirname or show_serie:
-                dirname = (dirname + "/" if dirname else '') + basename
-            path = (dirname + '/' if dirname else '') + (prefix if prefix else '') + (get_valid_filename(build.pretty_name()) if show_serie else '') + (type_str if type_str else '') + ('' if not suffix else ("-" + suffix)) + ext
-
-    folder = os.path.dirname(path)
-    if folder and not os.path.exists(folder):
-        os.makedirs(folder)
-    return path
-
-
-def build_output_filename(options, repo_list):
-    if options.graph_filename is None:
-        filename = 'compare/' + os.path.splitext(os.path.basename(options.test_files))[0] + '_' + '_'.join(
-            ["%s" % repo.reponame for repo in repo_list]) + '.pdf'
-    else:
-        filename = options.graph_filename
-    return filename
-
-def replace_path(path, build = None):
-    if build:
-        if build.version is not None:
-            path = path.replace('$version', build.version)
-        for var,val in build.repo.env.items():
-            path = path.replace('$'+var,val)
-    return path
-
-def parseBool(s):
-    if type(s) is str and s.lower() == "false":
-       return False
-    else:
-       return bool(s)
-
-def parseUnit(u):
-    r = re.match('([-]?)([0-9.]+)[ ]*([GMK]?)',u)
-    if r != None:
-        n = float(r.group(2))
-        unit = r.group(3)
-        if r.group(1) == "-":
-            n = -n
-
-        if unit is None or unit == '':
-            return n
-        if unit == 'G':
-            n = n * 1000000000
-        elif unit == 'M':
-            n = n * 1000000
-        elif unit == 'K':
-            n = n * 1000
-        else:
-            raise Exception('%s is not a valid unit !' % unit)
-        return n
-    else:
-        raise Exception("%s is not a number !" % u)
-
-def all_num(l):
-    for x in l:
-        if type(x) is not int and type(x) is not Decimal and not isinstance(x, (np.floating, float)):
-            return False
-    return True
-
-
-def ensure_folder_exists(filename):
-    savedir = Path(os.path.dirname(filename))
-    if not savedir.exists():
-        os.makedirs(savedir.as_posix())
-
-    if not os.path.isabs(filename):
-        filename = os.getcwd() + os.sep + filename
-    return filename
