+++ npf/output/__init__.py
@@ -1,55 +1,7 @@
 from typing import Dict, List, Tuple
-from npf import npf
-from npf.build import Build
-from npf.pipeline import pypost
-from npf.regression import Grapher, OrderedDict, Regression, npf
-from npf.repository import Repository
-from npf.statistics import Statistics
-from npf.test import Test
-from npf.types.dataset import Dataset, Run
 from npf.types.series import Series
 
-
-"""Runs all tests for a given list of tests (or a folder to expand), and a series of repositories.
-"""
-class Comparator():
-    def __init__(self, repo_list: List[Repository]):
-        self.repo_list = repo_list
-        self.graphs_series = []
-        self.time_graphs_series = []
-
-    def build_list(self, on_finish, test, build:Build, data_datasets:Dataset, time_datasets):
-         on_finish(self.graphs_series + [(test,build,data_datasets[0])], self.time_graphs_series + [(test,build,time_datasets[0])])
-
-    def run(self, test_name, options, tags:List, on_finish=None, do_regress=True):
-        for i_repo, repo in enumerate(self.repo_list):
-            build = None
-            regressor = Regression(repo)
-            tests = Test.expand_folder(test_name, options=options, tags=repo.tags + tags)
-            tests = npf.override(options, tests)
-            for test in tests:
-                build, data_dataset, time_dataset = regressor.regress_all_tests(
-                    tests=[test],
-                    options=options,
-                    do_compare=do_regress,
-                    on_finish=lambda b,dd,td: self.build_list(on_finish,test,b,dd,td) if on_finish else None,
-                    i_serie=i_repo,
-                    nseries=len(self.repo_list)
-                    )
-
-
-            if len(tests) > 0 and build is not None:
-                build._pretty_name = repo.name
-                self.graphs_series.append((test, build, data_dataset[0]))
-                self.time_graphs_series.append((test, build, time_dataset[0]))
-        if len(self.graphs_series) == 0:
-            print("No valid tags/test/repo combination.")
-            return None, None
-
-        return self.graphs_series, self.time_graphs_series
-
-
-def group_series(filename: str, series: Series , time_series:Series, options) -> Tuple[Dataset,Dict]:
+def generate_outputs(filename: str, series: Series , time_series:Series, options) -> Tuple['Dataset',Dict]:
     """
     The function merge different series together, finding common variables
 
@@ -63,6 +15,11 @@ def group_series(filename: str, series: Series , time_series:Series, options) ->
     the output. These options can include things like the file format, the delimiter to use, whether or
     not to include headers, etc
     """
+    import npf
+    from npf.output.statistics import Statistics
+    from npf.tests import pypost
+    from npf.tests.regression import Grapher, OrderedDict, npf
+
     if series is None:
         return
 
@@ -110,7 +67,7 @@ def group_series(filename: str, series: Series , time_series:Series, options) ->
                            dataset,
                            test,
                            max_depth=options.statistics_maxdepth,
-                           filename=options.statistics_filename or npf.build_output_filename(options, [build.repo for t,build,d in series]))
+                           filename=options.statistics_filename or npf.build_output_filename([build.repo for t,build,d in series]))
 
     common_variables = set.intersection(*map(set, all_variables))
