+++ npf/output/graph/variable_to_series.py
@@ -1,5 +1,6 @@
 import sys
-from npf.graph import Graph
+from npf.output.graph.graphdata import GraphData
+
 if sys.version_info < (3, 7):
     from orderedset import OrderedSet
 else:
@@ -7,7 +8,7 @@ else:
 from collections import OrderedDict
 
 # Extract the variable key so it becomes a serie
-def extract_variable_to_series(grapher, key, vars_values, all_results, dyns, build, script) -> Graph:
+def extract_variable_to_series(grapher, key, vars_values, all_results, dyns, build, script) -> GraphData:
     if not key in dyns:
         raise ValueError("Cannot extract %s because it is not a dynamic variable (%s are)" % (key, ', '.join(dyns)))
     dyns.remove(key)
@@ -21,7 +22,7 @@ def extract_variable_to_series(grapher, key, vars_values, all_results, dyns, bui
     except TypeError:
         print("ERROR : Cannot sort the following values :", values)
         return
-    new_varsall = OrderedSet()
+
     for i, value in enumerate(values):
         newserie = OrderedDict()
         for run, run_results in all_results.items():
@@ -31,7 +32,6 @@ def extract_variable_to_series(grapher, key, vars_values, all_results, dyns, bui
                 newrun = run.copy()
                 del newrun.variables[key]
                 newserie[newrun] = run_results
-                new_varsall.add(newrun)
 
         if type(value) is tuple:
             value = value[1]
@@ -42,7 +42,6 @@ def extract_variable_to_series(grapher, key, vars_values, all_results, dyns, bui
             nb._marker = grapher.graphmarkers[i % len(grapher.graphmarkers)]
         series.append((script, nb, newserie))
         grapher.glob_legend_title = grapher.var_name(key)
-    vars_all = list(new_varsall)
 
     if len(dyns) == 1:
         key = dyns[0]
@@ -53,12 +52,10 @@ def extract_variable_to_series(grapher, key, vars_values, all_results, dyns, bui
         key = "Variables"
         do_sort = False
     do_sort = grapher.config_bool_or_in('graph_x_sort', key, default=do_sort)
-    if (do_sort):
-        vars_all.sort()
-    graph = Graph(grapher)
+
+    graph = GraphData(grapher)
     graph.do_sort = do_sort
     graph.key = key
-    graph.vars_all = vars_all
     graph.vars_values = vars_values
     graph.series = series
     return graph
