+++ npf/output/statistics.py
@@ -3,6 +3,8 @@ import os
 import numpy as np
 import pydotplus as pydotplus
 import sys
+
+import npf.globals
 if sys.version_info < (3, 7):
     from orderedset import OrderedSet
 else:
@@ -13,10 +15,14 @@ from collections import OrderedDict
 
 from typing import List
 
-from npf.build import Build
-from npf.test import Test
-from npf.types.dataset import Dataset
-from npf import npf
+from npf.tests.build import Build
+from npf.tests.test import Test
+from npf.types.dataset import Dataset, sanitize
+import npf
+
+import pandas as pd
+import seaborn as sn
+import matplotlib.pyplot as plt
 
 class Statistics:
     @staticmethod
@@ -26,11 +32,25 @@ class Statistics:
         #Transform the dataset into a standard table of X/Features and Y observations
         dataset = Statistics.buildDataset(all_results, test)
 
-        #There's one per serie, so for each of those
+        # If we have too many metrics, we select the one with the most variance
+        all_vars = []
         for result_type, X, y, dtype in dataset:
+            all_vars.append(np.var(y)/np.mean(y))
+
+        if npf.globals.options.statistics_maxmetrics and len(dataset) > npf.globals.options.statistics_maxmetrics:
+            print(f"There are {len(dataset)} observations. Only the first {npf.globals.options.statistics_maxmetrics} with the most variance will be shown. Use --statistics-maxmetrics N to change the limit.")
+            s_dataset = [dataset[i] for i in np.flip(np.argsort(all_vars))[:npf.globals.options.statistics_maxmetrics]]
+        else:
+            s_dataset = dataset
+
+        #There's one per serie, so for each of those
+        for i_dataset, (result_type, X, y, dtype) in enumerate(s_dataset):
             if len(dataset) > 1:
-                print("Statistics for %s" % result_type)
+                print(f"Statistics for {result_type}")
+            if npf.globals.options.statistics_maxmetrics and i_dataset > npf.globals.options.statistics_maxmetrics:
+                break
             print("Learning dataset built with %d samples and %d features..." % (X.shape[0], X.shape[1]))
+            dataset = Statistics.buildDataset(all_results, test)
             clf = tree.DecisionTreeRegressor(max_depth=max_depth)
             try:
                 clf = clf.fit(X, y)
@@ -51,7 +71,7 @@ class Statistics:
                     f = npf.build_filename(test, build, filename if not filename is True else None, {}, 'pdf', result_type, show_serie=False, suffix="clf")
                     try:
                         graph.write(f, format=os.path.splitext(f)[1][1:])
-                        print("Decision tree visualization written to %s" % f)
+                        print(f"Decision tree visualization written to {f}")
                     except Exception as e:
                         print("Could not generate the tree vizualization : %s" % str(e))
 
@@ -60,7 +80,7 @@ class Statistics:
             print("")
             for i, column in enumerate(X.T):
                 varname = dtype['names'][i]
-                vars_values[varname] = set([v for v in np.unique(column)])
+                vars_values[varname] = set(list(np.unique(column)))
 
             print("")
             print("Feature importance:")
@@ -105,47 +125,20 @@ class Statistics:
                         print("  %s : None" % vs)
                     else:
                         print("  %s : %.02f " % (vs, tot / n))
-                print("")
+            print("")
 
-            print('')
-
-            ys = np.ndarray(shape = (len(X), len(dataset)))
-
-            for i,d in enumerate(dataset):
-                ys[:,i] = d[2]
-            import pandas as pd
-            df = pd.DataFrame(np.concatenate((X,ys),axis=1),columns=list(vars_values.keys()) + [d[0] if d[0] else "y" for d in dataset])
-            pd.options.display.float_format = "{:,.2f}".format
-            print("Correlation matrix:")
-
-            corr = df.corr()
-            corr = corr.dropna(axis=0,how='all')
-
-            corr = corr.dropna(axis=1,how='all')
-            keep = np.triu(np.ones(corr.shape)).astype(bool)
-            corr = corr.where(keep==True, np.nan)
-            print(corr.fillna(""))
-
-            import seaborn as sn
-            import matplotlib.pyplot as plt
-            ax = sn.heatmap(corr, cmap="viridis", fmt=".2f", annot=True)
-            ax.figure.tight_layout()
-            f = npf.build_filename(test, build, filename if not filename is True else None, {}, 'pdf', result_type, show_serie=False, suffix="correlation")
-            plt.savefig(f)
-            print(f"Graph of correlation matrix saved to {f}")
-
-            plt.clf()
-
+            #P-value of ANOVA
             import statsmodels.api as sm
-
             from statsmodels.formula.api import ols
             variables = [k for k,v in vars_values.items() if len(v) > 1]
 
-            m = result_type + ' ~ ' +' * '.join( [f"{v}" for v in variables])
+            m = f'{result_type} ~ ' + ' * '.join( [f"{v}" for v in variables])
 
             pd.set_option("display.max_rows", None)
 
-            dfn = df[variables + [result_type]].replace([np.inf, -np.inf], np.nan).dropna(axis=1)
+            d = np.concatenate((X,y.reshape(-1,1)),axis=1)
+
+            dfn = pd.DataFrame(d, columns=list(vars_values.keys()) + [result_type or "y"])
             dfn = dfn.replace([np.inf, -np.inf], np.nan).dropna(axis=0)
 
             model =  ols(m, data=dfn).fit()
@@ -178,6 +171,34 @@ class Statistics:
 
             pd.reset_option("display.float_format")
 
+        print('')
+
+        ys = np.ndarray(shape = (len(X), len(dataset)))
+
+        for i,d in enumerate(dataset):
+            ys[:,i] = d[2]
+
+        df = pd.DataFrame(np.concatenate((X,ys),axis=1),columns=list(vars_values.keys()) + [d[0] if d[0] else "y" for d in dataset])
+        pd.options.display.float_format = "{:,.2f}".format
+        print("Correlation matrix:")
+
+        corr = df.corr()
+        corr = corr.dropna(axis=0,how='all')
+
+        corr = corr.dropna(axis=1,how='all')
+        keep = np.triu(np.ones(corr.shape)).astype(bool)
+        corr = corr.where(keep==True, np.nan)
+        print(corr.fillna(""))
+
+
+        ax = sn.heatmap(corr, cmap="viridis", fmt=".2f", annot=True)
+        ax.figure.tight_layout()
+        f = npf.build_filename(test, build, filename if not filename is True else None, {}, 'pdf', result_type, show_serie=False, suffix="correlation")
+        plt.savefig(f)
+        print(f"Graph of correlation matrix saved to {f}")
+
+        plt.clf()
+
     @classmethod
     def buildDataset(cls, all_results: Dataset, test: Test) -> List[tuple]:
         #map of every <variable name, format>
@@ -186,7 +207,7 @@ class Statistics:
         y = OrderedDict()
         dataset = []
         for i, (run, results_types) in enumerate(all_results.items()):
-            vars = list(run.read_variables()[k] for k in dtype['names'])
+            vars = list(run.read_variables()[k] for k in dtype['names'] if k in run.read_variables())
             if not results_types is None and len(results_types) > 0:
 
                 dataset.append([v for v in vars])
@@ -204,8 +225,10 @@ class Statistics:
                     row[i] = values.index(row[i])
                 dtype['values'][i] = list(values)
         X = np.array(dataset, ndmin=2)
-
+        
         lset = []
-        for result_type, v in y.items():
-            lset.append((result_type, X, np.array(v),dtype))
+        for result_type, y_value in y.items():
+            result_type = sanitize(result_type)
+            lset.append((result_type, X, np.array(y_value),dtype))
+
         return lset
