+++ npf/types/dataset.py
@@ -11,10 +11,8 @@ else:
 import natsort
 import csv
 
-from npf import npf
-from npf.variable import is_numeric, get_numeric, numeric_dict
-
-from npf.types.web.web import prepare_web_export
+from npf.tests.variable import is_numeric, get_numeric, numeric_dict
+from npf.output.web.web import prepare_web_export
 
 class Run:
     def __init__(self, variables):
@@ -476,3 +474,5 @@ def convert_to_xyeb(datasets: List[Tuple['Test', 'Build' , Dataset]], run_list, 
     return data_types
 
 
+def sanitize(name:str):
+    return name.replace('-', '_')
